Created single backHome button and updated home, projects, contact, and about me pages to include new button component.
