+++ src/components/AnimationRoutes/AnimationRoutes.jsx
@@ -15,7 +15,6 @@ export default function AnimationRoutes() {
       <Routes key={location.pathname} location={location}>
         <Route path="/" element={<HomePage />}></Route>
         <Route path="/about-me" element={<AboutMePage />}></Route>
-
         <Route path="/project-cards" element={<ProjectCardsPage />}></Route>
         <Route path="/project-page" element={<ProjectPage />}></Route>
         <Route path="/resume" element={<ResumePage />}></Route>
